+++ src/reconstructed/map_activation.cpp
@@ -25,7 +25,7 @@ public:
     InitCallback(const KinokoSquirrelObject *environment, const char *name) {
         kinoko_sqplus_object_get_value((void *)(environment), (void *)(&object_), name);
     }
-    ~InitCallback() { kinoko_sqplus_object_destroy((void *)(&object_)); }
+    ~InitCallback() { (int32_t)(intptr_t)(kinoko_sqplus_object_destroy((void *)(&object_))); }
     InitCallback(const InitCallback&) = delete;
     InitCallback& operator=(const InitCallback&) = delete;
     bool is_closure() const { return object_.value._type == OT_CLOSURE; }
